backend/requirements.txt
# Add chess engine library for robust USI communication
fastapi
uvicorn
python-shogi
google-generativeai
python-dotenv
